pyproject.toml
[tool.poetry]
name = "tumpara"
version = "0.1.0"
description = "Server component of Tumpara, a self-hosted solution for managing home media libraries."
authors = ["Yannik Rödel <hey@yannik.info>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
blurhash-python = "^1.1.1"
django-cors-headers = "^3.7.0"
graphene-django = "^2.15.0"
graphene-gis = "^0.0.6"
inotifyrecursive = "^0.3.5"
Pillow = "^8.2.0"
psycopg2 = "^2.8.6"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.8.0"
freezegun = "^1.1.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
PyYAML = "^5.4.1"
hypothesis = "^6.14.0"
selenium = "^3.141.0"
pygments-graphql = {version = "^1.0.0", extras = ["docs"]}
Sphinx = {version = "^4.0.2", extras = ["docs"]}
sphinx-rtd-theme = {version = "^0.5.2", extras = ["docs"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
